fix rmtree system_theme_dir don't exists bug

+++ modules/individuation/src/xdg_support.py
@@ -112,7 +112,10 @@ def get_system_theme_dir():
     program_theme_dir = os.path.join(program_dir, "theme")
 
     system_theme_dir = os.path.join(get_config_dir(), "system-theme")
-    shutil.rmtree(system_theme_dir)
+    if os.path.exists(system_theme_dir):
+        try:
+            shutil.rmtree(system_theme_dir)
+        except: pass    
     os.makedirs(system_theme_dir)
     for theme_file in os.listdir(program_theme_dir):
         shutil.copy("%s/%s" % (program_theme_dir, theme_file), "%s/%s" % (system_theme_dir , theme_file))
